create secrets utils
