+++ liminal/runners/airflow/executors/kubernetes.py
@@ -66,6 +66,7 @@ class KubernetesPodExecutor(executor.Executor):
 
     def _volumes(self):
         volumes_config = self.liminal_config.get('volumes', [])
+        secrets_config = self.liminal_config.get('secrets', [])
         volumes = []
         for volume_config in volumes_config:
             name = volume_config['volume']
@@ -78,6 +79,12 @@ class KubernetesPodExecutor(executor.Executor):
                 volume = V1Volume(name=name, persistent_volume_claim={'claimName': claim_name})
 
             volumes.append(volume)
+
+        for secret_config in secrets_config:
+            name = secret_config['secret']
+            secret = V1Volume(name=name, secret={'secretName': name})
+            volumes.append(secret)
+
         return volumes
 
     def __kubernetes_kwargs(self, task: ContainerTask):
